internal/handler/order_handler.go
package handler

import (
	"OrderService/internal/cache"
	"encoding/json"
	"net/http"

	"github.com/go-chi/chi"
)

func OrderHandler(c cache.OrderStore) http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		orderUID := chi.URLParam(r, "order_uid")
		if orderUID == "" {
			http.Error(w, "order_uid is required", http.StatusBadRequest)
			return
		}
		order, found := c.Get(orderUID)
		if !found {
			http.Error(w, "order not found", http.StatusNotFound)
			return
		}
		w.Header().Set("Content-Type", "application/json")
		if err := json.NewEncoder(w).Encode(order); err != nil {
			http.Error(w, "Failed to encode response", http.StatusInternalServerError)
		}
	}
}
